made it easier to request a session using discord
I split the session method into 3, a board, a request and a helper. Because of this users can now use tab-completion on the command completion. This is faster and eleminates user error, since the user doesn't have to enter an exact string. also added some unit-tests for the session-file. The file was to big and complex to test all at once.